requirements.txt
appdirs==1.4.3
Babel==2.4.0
click==6.7
ecdsa==0.13
Flask==0.12
Flask-Babel==0.11.1
Flask-Table==0.4.1
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pycrypto==2.6.1
PyGithub==1.32
pyparsing==2.2.0
python-jose==1.3.2
pytz==2016.10
six==1.10.0
Werkzeug==0.12.1
